Split DatabaseInterface and internal ProcessIoDatabase implementation
This is done to achieve better separation of concerns in preparation for
creating different database communication implementations for Windows
support in the future.
