+++ examples/insert.php
@@ -1,6 +1,6 @@
 <?php
 
-use Clue\React\SQLite\Database;
+use Clue\React\SQLite\DatabaseInterface;
 use Clue\React\SQLite\Factory;
 use Clue\React\SQLite\Result;
 
@@ -10,7 +10,7 @@ $loop = React\EventLoop\Factory::create();
 $factory = new Factory($loop);
 
 $n = isset($argv[1]) ? $argv[1] : 1;
-$factory->open('test.db')->then(function (Database $db) use ($n) {
+$factory->open('test.db')->then(function (DatabaseInterface $db) use ($n) {
     $db->exec('CREATE TABLE IF NOT EXISTS foo (id INTEGER PRIMARY KEY AUTOINCREMENT, bar STRING)');
 
     for ($i = 0; $i < $n; ++$i) {
