+++ tests/FunctionalDatabaseTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use Clue\React\SQLite\Database;
+use Clue\React\SQLite\DatabaseInterface;
 use Clue\React\SQLite\Factory;
 use Clue\React\SQLite\Result;
 use PHPUnit\Framework\TestCase;
@@ -15,10 +15,10 @@ class FunctionalDatabaseTest extends TestCase
         $promise = $factory->open(':memory:');
 
         $promise->then(
-            $this->expectCallableOnceWith($this->isInstanceOf('Clue\React\SQLite\Database'))
+            $this->expectCallableOnceWith($this->isInstanceOf('Clue\React\SQLite\DatabaseInterface'))
         );
 
-        $promise->then(function (Database $db) {
+        $promise->then(function (DatabaseInterface $db) {
             $db->close();
         });
 
@@ -33,10 +33,10 @@ class FunctionalDatabaseTest extends TestCase
         $promise = $factory->open(':memory:');
 
         $promise->then(
-            $this->expectCallableOnceWith($this->isInstanceOf('Clue\React\SQLite\Database'))
+            $this->expectCallableOnceWith($this->isInstanceOf('Clue\React\SQLite\DatabaseInterface'))
         );
 
-        $promise->then(function (Database $db) {
+        $promise->then(function (DatabaseInterface $db) {
             $db->quit();
         });
 
@@ -64,7 +64,7 @@ class FunctionalDatabaseTest extends TestCase
         $this->assertTrue(is_resource($server));
         fclose($server);
 
-        $promise->then(function (Database $db) {
+        $promise->then(function (DatabaseInterface $db) {
             $db->close();
         });
 
@@ -109,7 +109,7 @@ class FunctionalDatabaseTest extends TestCase
         $promise = $factory->open(':memory:');
 
         $once = $this->expectCallableOnce();
-        $promise->then(function (Database $db) use ($once){
+        $promise->then(function (DatabaseInterface $db) use ($once){
             $db->quit()->then($once);
         });
 
@@ -130,7 +130,7 @@ class FunctionalDatabaseTest extends TestCase
         $promise = $factory->open(':memory:');
 
         $once = $this->expectCallableOnce();
-        $promise->then(function (Database $db) use ($once){
+        $promise->then(function (DatabaseInterface $db) use ($once){
             $db->quit()->then($once);
         });
 
@@ -149,7 +149,7 @@ class FunctionalDatabaseTest extends TestCase
         $promise = $factory->open(':memory:');
 
         $once = $this->expectCallableOnce();
-        $promise->then(function (Database $db) use ($once){
+        $promise->then(function (DatabaseInterface $db) use ($once){
             $db->quit();
             $db->quit()->then(null, $once);
         });
@@ -165,7 +165,7 @@ class FunctionalDatabaseTest extends TestCase
         $promise = $factory->open(':memory:');
 
         $data = null;
-        $promise->then(function (Database $db) use (&$data){
+        $promise->then(function (DatabaseInterface $db) use (&$data){
             $db->query('SELECT 1 AS value')->then(function (Result $result) use (&$data) {
                 $data = $result->rows;
             });
@@ -186,7 +186,7 @@ class FunctionalDatabaseTest extends TestCase
         $promise = $factory->open(':memory:');
 
         $data = null;
-        $promise->then(function (Database $db) use (&$data){
+        $promise->then(function (DatabaseInterface $db) use (&$data){
             $db->query('SELECT "hellö" AS value')->then(function (Result $result) use (&$data) {
                 $data = $result->rows;
             });
@@ -207,7 +207,7 @@ class FunctionalDatabaseTest extends TestCase
         $promise = $factory->open(':memory:');
 
         $data = null;
-        $promise->then(function (Database $db) use (&$data){
+        $promise->then(function (DatabaseInterface $db) use (&$data){
             $db->query('SELECT ? AS value', array(1))->then(function (Result $result) use (&$data) {
                 $data = $result->rows;
             });
@@ -228,7 +228,7 @@ class FunctionalDatabaseTest extends TestCase
         $promise = $factory->open(':memory:');
 
         $data = null;
-        $promise->then(function (Database $db) use (&$data){
+        $promise->then(function (DatabaseInterface $db) use (&$data){
             $db->query('SELECT :value AS value', array('value' => 1))->then(function (Result $result) use (&$data) {
                 $data = $result->rows;
             });
@@ -249,7 +249,7 @@ class FunctionalDatabaseTest extends TestCase
         $promise = $factory->open(':memory:');
 
         $data = null;
-        $promise->then(function (Database $db) use (&$data){
+        $promise->then(function (DatabaseInterface $db) use (&$data){
             $db->query('SELECT ? AS value', array(null))->then(function (Result $result) use (&$data) {
                 $data = $result->rows;
             });
@@ -270,7 +270,7 @@ class FunctionalDatabaseTest extends TestCase
         $promise = $factory->open(':memory:');
 
         $once = $this->expectCallableOnceWith($this->isInstanceOf('RuntimeException'));
-        $promise->then(function (Database $db) use ($once){
+        $promise->then(function (DatabaseInterface $db) use ($once){
             $db->query('nope')->then(null, $once);
 
             $db->quit();
@@ -287,7 +287,7 @@ class FunctionalDatabaseTest extends TestCase
         $promise = $factory->open(':memory:');
 
         $once = $this->expectCallableOnceWith($this->isInstanceOf('RuntimeException'));
-        $promise->then(function (Database $db) use ($once){
+        $promise->then(function (DatabaseInterface $db) use ($once){
             $db->query('SELECT 1')->then(null, $once);
 
             $db->close();
@@ -304,7 +304,7 @@ class FunctionalDatabaseTest extends TestCase
         $promise = $factory->open(':memory:');
 
         $data = 'n/a';
-        $promise->then(function (Database $db) use (&$data){
+        $promise->then(function (DatabaseInterface $db) use (&$data){
             $db->exec('CREATE TABLE foo (bar STRING)')->then(function (Result $result) use (&$data) {
                 $data = $result->rows;
             });
@@ -325,7 +325,7 @@ class FunctionalDatabaseTest extends TestCase
         $promise = $factory->open(':memory:');
 
         $once = $this->expectCallableOnceWith($this->isInstanceOf('RuntimeException'));
-        $promise->then(function (Database $db) use ($once){
+        $promise->then(function (DatabaseInterface $db) use ($once){
             $db->exec('USE a')->then(null, $once);
 
             $db->close();
@@ -342,7 +342,7 @@ class FunctionalDatabaseTest extends TestCase
         $promise = $factory->open(':memory:');
 
         $once = $this->expectCallableOnceWith($this->isInstanceOf('RuntimeException'));
-        $promise->then(function (Database $db) use ($once){
+        $promise->then(function (DatabaseInterface $db) use ($once){
             $db->close();
             $db->exec('USE a')->then(null, $once);
         });
